four
The "About" page should contain a short paragraph about the site.

The "Contact" page should contain a form with the following elements:
- textfield for Name (required)
- texfield for email (required and validated as email)
- checkboxes for "Category interests" (User can select multiple from "Industry","Technical" or "Career")
- a select box for "Your role" (User can select only one from "Writer","Contributor" or "Administrator")
- A submit button
The form should submit to a "process-contact.php" using the POST method. The code for this page is not necessary for this assignment

This is NOT a signup or registration form. It is a contact form.

Add a favicon and appropriate metadata to your pages.